Lorentz.Contracts.ManagedLedger.Babylon

Description

Managed ledger which is compatible with FA1.2 standard and extended with administrator functionality. It is not compatible with the Athens version of FA1.2.

Documentation

data Parameter

Constructors

Transfer !TransferParams

Approve !ApproveParams

GetAllowance !(View GetAllowanceParams Natural)

GetBalance !(View Address Natural)

GetTotalSupply !(View () Natural)

SetPause !Bool

SetAdministrator !Address

GetAdministrator !(View () Address)

Mint !MintParams

Burn !BurnParams

type Storage = Storage' StorageFields

data Error

Constructors

UnsafeAllowanceChange Natural
Attempt to change allowance from non-zero to a non-zero value.

SenderIsNotAdmin
Contract initiator has not enough rights to perform this operation.

NotEnoughBalance ("required" :! Natural, "present" :! Natural)
Insufficient balance.

NotEnoughAllowance ("required" :! Natural, "present" :! Natural)
Insufficient allowance to transfer foreign funds.

OperationsArePaused
Operation is unavailable until resume by token admin.

mkStorage :: Address -> Map Address Natural -> Storage

Create a default storage with ability to set some balances to non-zero values.

managedLedgerContract :: Contract Parameter Storage

module Lorentz.Contracts.ManagedLedger.Types