Lorentz.Contracts.ManagedLedger.Impl

Description

Implementation of manalged ledger which does not require particular storage type.

Documentation

type StorageFieldsC fields = fields `HasFieldsOfType` ["admin" := Address, "paused" := Bool, "totalSupply" := Natural]

transfer :: forall fields. StorageFieldsC fields => Entrypoint TransferParams fields

transfer' :: forall fields. StorageFieldsC fields => Entrypoint ("sender" :! Address, TransferParams) fields

This is a slightly more general version of transfer which is needed to implement transferViaProxy. It does not use sender.

approve :: forall fields. StorageFieldsC fields => Entrypoint ApproveParams fields

approve' :: forall fields. StorageFieldsC fields => Entrypoint AllowanceParams fields

This is a slightly more general version of approve which is needed to implement approveViaProxy. It does not use sender.

getAllowance :: StorageFieldsC fields => Entrypoint (View GetAllowanceParams Natural) fields

getBalance :: StorageFieldsC fields => Entrypoint (View Address Natural) fields

getTotalSupply :: StorageFieldsC fields => Entrypoint (View () Natural) fields

setPause :: StorageFieldsC fields => Entrypoint Bool fields

setAdministrator :: StorageFieldsC fields => Entrypoint Address fields

getAdministrator :: StorageFieldsC fields => Entrypoint (View () Address) fields

mint :: StorageFieldsC fields => Entrypoint MintParams fields

burn :: StorageFieldsC fields => Entrypoint BurnParams fields