Lorentz.Contracts.ManagedLedger.Types

Description

Types shared between contracts participating in ManagedLedger.

Documentation

type TransferParams = ("from" :! Address, "to" :! Address, "value" :! Natural)

type ApproveParams = ("spender" :! Address, "value" :! Natural)

type AllowanceParams = ("owner" :! Address, "spender" :! Address, "value" :! Natural)

type GetAllowanceParams = ("owner" :! Address, "spender" :! Address)

type MintParams = ("to" :! Address, "value" :! Natural)

type BurnParams = ("from" :! Address, "value" :! Natural)

type LedgerValue = ("balance" :! Natural, "approvals" :! Map Address Natural)

data Error

Constructors

UnsafeAllowanceChange Natural
Attempt to change allowance from non-zero to a non-zero value.

SenderIsNotAdmin
Contract initiator has not enough rights to perform this operation.

NotEnoughBalance ("required" :! Natural, "present" :! Natural)
Insufficient balance.

NotEnoughAllowance ("required" :! Natural, "present" :! Natural)
Insufficient allowance to transfer foreign funds.

OperationsArePaused
Operation is unavailable until resume by token admin.

data Storage' fields

Constructors

Storage'

Fields

ledger :: BigMap Address LedgerValue

fields :: fields

mkStorage' :: Map Address Natural -> fields -> Storage' fields

Create a default storage with ability to set some balances to non-zero values.