Lorentz.Contracts.UpgradeableCounter

Description

UpgradeableCounter demonstrates the implementation of a simple contract that has upgradeable storage, interface, and implementation.

In the first version it stores a Natural and allows to add some value to it or multiply the current value by a certain natural number.

The second version changes the type of the stored value to Integer, and instead of providing Mul Natural and Add Natural endpoints, it just allows to increment or decrement the current value.

While the contract does not have any advanced functionality, it provides a birds-eye view on all the aspects of the upgradeable contracts concept and serves as an example on how to apply this concept.

Documentation

data Parameter (interface :: [EntryPointKind])

Constructors

Run (UParam interface)

Upgrade UpgradeParameters

GetVersion (View () Natural)

SetAdministrator Address

EpwBeginUpgrade Natural

EpwApplyMigration MigrationScript

EpwSetCode ContractCode

EpwFinishUpgrade

data Storage

data Error

Constructors

SenderIsNotAdmin
Tx sender does not have enough rights to perform this operation.

VersionMismatch ("expected" :! Natural, "actual" :! Natural)
Expected version does not match the version of the supplied code.

ContractIsPaused
The reuested operation requires the contract to be running but it is paused.

ContractIsNotPaused
The reuested operation requires the contract to be paused but it is not.

upgradeableCounterContract :: forall (interface :: [EntryPointKind]). Contract (Parameter interface) Storage

mkEmptyStorage :: Address -> Storage