Lorentz.Contracts.UpgradeableCounter.V1

Documentation

type Interface = ["add" ?: Natural, "mul" ?: Natural, "getCounterValue" ?: View () Natural]

version :: Natural

migrate :: MigrationScript

migrations :: [MigrationScript]

Migrations represent entrypoint-wise upgrades. Each migration puts an implementation of a method to UStore. The contract code itself (epwServe) does not do anything special except for taking these lambdas out of the big map.

counterContract :: ContractCode

type UStoreV1 = UStore UStoreTemplate