Lorentz.Contracts.UpgradeableUnsafeLedger

Description

UpgradeableUnsafeLedger demonstrates a real-world use case of an upgradeable contract. Its V1 contains a bug that makes it return an incorrect totalSupply. Fortunately, we can upgrade the contract and fix the bug. This contract does not use entrypoint-wise upgrades. Instead, it upgrades atomically in one transaction. For non-atomic upgrades please see the UpgradeableCounter example.

Documentation

data Parameter (interface :: [EntryPointKind])

Constructors

Run (UParam interface)

Upgrade UpgradeParameters

GetVersion (View () Natural)

SetAdministrator Address

EpwBeginUpgrade Natural

EpwApplyMigration MigrationScript

EpwSetCode ContractCode

EpwFinishUpgrade

data Storage

upgradeableUnsafeLedgerContract :: forall (interface :: [EntryPointKind]). Contract (Parameter interface) Storage

mkEmptyStorage :: Address -> Storage